+++ RobWork/src/rwlibs/simulation/SimulatedScanner25D.cpp
@@ -56,7 +56,7 @@ public:
 SimulatedScanner25D::SimulatedScanner25D(const std::string& name,
                                          rw::kinematics::Frame *frame,
                                          FrameGrabber25D::Ptr framegrabber):
-		SimulatedSensor( rw::common::ownedPtr( new Scanner25DModel(name,framegrabber->getWidth(), framegrabber->getHeight(), frame) )),
+		SimulatedSensor( rw::common::ownedPtr( new Scanner25DModel(name,static_cast<int>(framegrabber->getWidth()), static_cast<int>(framegrabber->getHeight()), frame) )),
 		_framegrabber(framegrabber),
 		_frameRate(30),
         _dtsum(0)
@@ -68,7 +68,7 @@ SimulatedScanner25D::SimulatedScanner25D(const std::string& name,
 		const std::string& desc,
 		rw::kinematics::Frame *frame,
 		FrameGrabber25D::Ptr framegrabber):
-		SimulatedSensor( rw::common::ownedPtr( new Scanner25DModel(name,framegrabber->getWidth(), framegrabber->getHeight(), frame) )),
+		SimulatedSensor( rw::common::ownedPtr( new Scanner25DModel(name,static_cast<int>(framegrabber->getWidth()), static_cast<int>(framegrabber->getHeight()), frame) )),
 		_framegrabber(framegrabber),
 		_frameRate(30),
 		_dtsum(0)
